Extract postMessage listener to preview.ts utils
Moves isInIframe and listenForPreviewMessages() out of +layout.svelte
into a dedicated utils file for better separation of concerns.
